Don't trigger TLS handshake until first I/O operation. (#1106)
Enables read timeouts to work correctly with OpenJDK's
HttpsURLConnection. Should be a no-op on Android as it uses a
different HttpsURLConnection implementation but there is some
minor app compat risk (see below).

With OpenJDK's HttpsURLConnection, the order of operations in the
internal Sun classes is essentially:

* HttpClient
  * Makes plain socket connection using connect timeout
  * Stashes the OutputStream from the socket using getOutputStream()
  * Calls afterConnect() which is overridden in HttpsClient
* HttpsClient.afterConnect()
  * Sets socket read timeout to HttpsURLConnection's read timeout value
  * Triggers TLS handsshake via startHandshake()

This works fine on OpenJDK, but with Conscrypt the TLS handshake
is triggered by the getOutputStream() call and so proceeds before
a read timeout has been set and can potentially hang forever, and
does in some Cloud apps.

There are some subtle concurrency issues and missed optimisations
in the Conscrypt code, so this change does the bare minimum to
re-order operations and I'll address the other issues in followup
changes.

The socket changes are:
* Do not trigger a handshake from the Stream constructors.
* In the stream I/O methods, call waitForHandshake() rather
than startHandshake().  startHandshake() actually makes no
sense here, but it has worked up until now because of the
waitForHandshake() in the constructor and because in the
case where handshaking hasn't already started, startHandshake()
and doHandshake() are synchronous.

The rest of the change is test code - the only way to test
is to have an HTTP server to talk to which can inject faults
and I didn't really want to pull in anything as large as the
okhttp MockWebServer.

App compat risk: Handshake exceptions which were previously
thrown from get*Stream() may now be thrown from the read/write
methods.  This won't affect any app that explcitly calls
startHandshake() or where the getStream() and IO happen in
equivalent try/catch blocks.

Tested on Android: CtsLibcoreOkHttpTestCases MtsConscryptTestCases