Update build tools.
Much yak shaving:
* BoringSSL HEAD requires a recent-ish cmake.
* Conscrypt was using Android Gradle plugiun 4.x which defaults to cmake 3.10 and crashes with later versions
* Upgrading to plugin version 7.x requires upgrading to Gradle 7.x

This is a minimum viable upgrade to Gradle 7.5 plus plugin 7.4,
i.e. just enough to build and remove most warnings. Some sub-projects
are broken, but I believe they were broken before (e.g. conscrypt-android-platform)
and there is still much scope for optimisation.

This update is orthogonal to the ARM changes in #1100 and once
this lands I'll rebase that PR on top of it.

Once #1100 lands we can consider Gradle 8.  :)