Tidy ConscryptEngineSocket state machine. (#1120)
Previously the socket would never actually reach STATE_READY
but it didn't matter.

This tidy-up makes the states and transitions more explicit
and moves TLS metric logging to the correct place to prevent
duplicate logging if waitForHandshake is called multiple times.

Also moves calling of the SSLSocket handshake listeners outside
of the SSLEngine handshake callback, this means they can
operate on the socket without causing re-entrant calls to wrap()
etc.  This behaviour isn't defined to be allowed anywhere and
it doesn't work with the FD socket.

Finally tidied a bunch of IDEA warnings from SSLSocketTest and
used assertThrows where possible.