FdSocket: Count TLS handshake stats during state transitions. (#1125)
This is analogous to #1120 and removes double-counting by only
counting metrics during explicit state changes.  There is a lot
more tidying up that could be done here, but as this class is
long deprecated, this change is kept to the minimum required to
ensure consistent counting between the EngineSocket and FdSocket.

Also incorporates the "failed" values from #1123 to make analysis
clearer.

TODO(prb): Tested manually: We have no real tests to ensure
consistent stats counting.