Use accessors for BIO and BIO_METHOD (#1134)
In doing so, this removes the puts callback. BIO_puts in BoringSSL is
implemented with BIO_write. We ignore the puts callback. I've also
removed the create callback and deferred BIO_set_init to when the stream
is attached. This makes no real difference, but since the BIO isn't
usable until the stream is attached, BIO_set_init probably should be
called later.