Deal with partially constructed objects during finalize(). (#1157)
The happy path for shutting down a ConscryptEngine is one of the
close* methods which then calls closeAndFreeResources()
synchronised on ssl, so all the logic can assume ssl is non-null.

However it also seems to be possible for ssl to be null when
finalizing, presumably if newSsl() threw, e.g. due to low memory.
If this is the case then we don't need to do anything as e.g.
there cannot be any session snapshot which needs saving.

And if ssl is non-null then the finalizer should also synchronise
of ssl for thread safety... *shouldn't* be necessary but does no
harm.