Reset HMAC and CMAC contexts more efficiently
If computing a new MAC with the same key, we would ideally save a little
bit of work. (A couple blocks of the underlying hash for HMAC and key
schedule calculation with CMAC.) Use the APIs for doing this, and also
cover this case with tests.

I noticed that MacSpi also has an optional clone() method, which we
could support here too, but I've left that for another PR.