Keep LogInfo in VerifiedSCT
There are 2 phases for certificate transparency validation:
1. Making sure that the SCTs are valid (i.e., from a known log, with a
   valid signature).
2. Making sure that the SCTs comply with the policy.

If valid, the LogInfo object is matched at step 1. For the policy
validation, some attributes of the LogInfo are required (i.e., the
operator name). Instead of looking up the LogInfo in step 2, keep a
reference of the LogInfo from step 1 in the VerifiedSCT object.

Bug: 319829948
Test: atest
Change-Id: Ifc1cd06d3ccb2d223777de7c5ca5d617e89ec2f1