Implement Android CT Policy for OCSP and TLS Data SCTs
This is the second part of the policy implementation which verifies SCTs
that are included in OCSP or TLS Data. The first part for embedded SCTs
was implemented in commit f70dd5ce7e40d14e361a0def773ced67e69c01f5.

The policy is based on the current Chrome policy[1].

[1] https://googlechrome.github.io/CertificateTransparency/ct_policy.html

Test: atest ConscryptPrivateTestCases
Bug: 319829948
Change-Id: If8b3959897559517ba8e447463b3d3f9f75322c9