Upstream AOSP (#1244)
* Add metric reporting for log list status

For the Android platform, report to statsd when a log list has been
loaded (successfully or not).

A new interface StatsLog is added to abstract ConscryptStatsLog (which
is renamed to StatsLogImpl). This interface is used for dependency
injection in LogStoreImpl.

The Platform class is marked as public so that LogStoreImpl (which
belongs to the ct subpackage) can access it.

The metrics methods (countTlsHandshake and the new
updateCTLogListStatusChanged) are moved out of Platform into the new
StatsLog. Platform is updated to simply return the implementation
instance (singleton).

Test: adb shell cmd stats print-logs; run ctdemo manually; confirm that
      the atom is sent to statsd correctly (via logcat)
Test: atest
Bug: 356111598
Bug: 319829948
Change-Id: I9f2709ef6c1acf349d197560c7580e8bf43c5281

* Implement Android CT Policy for OCSP and TLS Data SCTs

This is the second part of the policy implementation which verifies SCTs
that are included in OCSP or TLS Data. The first part for embedded SCTs
was implemented in commit f70dd5ce7e40d14e361a0def773ced67e69c01f5.

The policy is based on the current Chrome policy[1].

[1] https://googlechrome.github.io/CertificateTransparency/ct_policy.html

Test: atest ConscryptPrivateTestCases
Bug: 319829948
Change-Id: If8b3959897559517ba8e447463b3d3f9f75322c9


* Missing TargetApi annotation.

---------

Co-authored-by: Thiébaud Weksteen <tweek@google.com>