Report CT verification result
Add new stats method (reportCTVerificationResult) to report on the
outcome of CT verification. This includes reporting if CT verification
was requested but no log list was found.

A new platform method is added (reasonCTVerificationRequired) to capture
why the verification was triggered.

Note that reporting is only enabled if isCTVerificationRequired
returns true. That is, no metric is emitted unless an app opted-in to
Certificate Transparency.

Bug: 356111598
Test: atest
Test: adb shell cmd stats print-logs 1 && manually run test demo:
  - Without log list
  - With log list and domain providing SCTs
  - With log list and domain not providing SCTs
  All atoms are reported as expected.
Flag: com.android.org.conscrypt.flags.certificate_transparency_platform
Change-Id: Ibf93147e1014fafb44a52098f4552418b9a61850