Add support for enabling/disabling TLS v1.0 and 1.1 in Conscrypt
Note: this deprecates those protocols, but please let me know if I
should revert that part of the change.

This reverts commit 5ae7b5c2f272365d13c9ff3ba0c7a682375dc1d8.

Reason for revert: fixed the failures by checking that API levels are not higher than a 100, and bumped API level check to 36 because I noticed that this version of using reflection is more resilient than the previous one.

Change-Id: I42fcb922e046072eea0fa5aee07c513233d2b1e9