Fix zero size reads for Engine Socket.
InputStream explicitly allows zero size reads without attempting
IO or blocking.  We do this correctly for the fd-socket but
missed it for the Engine socket.

Not clear whether zero sized read should trigger a TLS handshake,
but it always has on Android so keeping it that way.

Bug manifested as a flaky test hang.  Fixed the test to
explicitly check corner cases and removed un-needed concurrency.