Use Android Gradle Plugin as a plugin
The Android Gradle Plugin was used as a dependency in the build script.
Simply apply the plugin as per Gradle's best practices:

"Instead of manipulating the script classpath directly, it is
recommended to apply plugins that come with their own classpath." [1]

Move the logic to detect if Android modules should be enabled to
settings.gradle.

[1] https://docs.gradle.org/7.5/userguide/tutorial_using_tasks.html#sec:build_script_external_dependencies