Merge metrics changes from AOSP (#1288)
* Decouple TrustManagerImpl from CT classes

TrustManagerImpl had multiple references to Certificate Transparency
objects: LogStore, Verifier, Policy, etc. Abstract these into a
unique class (ct.CertificateTransparency).

One constructor for TrustManagerImpl (4 parameters) is removed as it is
not used and not part of the API. The largest constructor is updated to
simply receive an ct.SubSystem instance.

This change is a no-op. It will help testing the logic of checkCT.

Bug: 356111598
Test: atest CtsLibcoreTestCases
Test: atest ConscryptPrivateTestCases
Change-Id: I03c5fd94aa5912035b54bb0912167f9e421de33b

* Remove LogStore.setPolicy

LogStore requires access to the policy when loading, to determine if the
log list is considered compliant or not. The policy was previously
associated using a separate setter (setPolicy). Force the passing of the
policy to the constructor of LogStore to avoid potentially missing that
call.

This is a no-op change.

Bug: 356111598
Test: atest ConscryptPrivateTestCases
Change-Id: Icfb5e3673d357abb59073509b976c5be342bdf08

* Report CT verification result

Add new stats method (reportCTVerificationResult) to report on the
outcome of CT verification. This includes reporting if CT verification
was requested but no log list was found.

A new platform method is added (reasonCTVerificationRequired) to capture
why the verification was triggered.

Note that reporting is only enabled if isCTVerificationRequired
returns true. That is, no metric is emitted unless an app opted-in to
Certificate Transparency.

Bug: 356111598
Test: atest
Test: adb shell cmd stats print-logs 1 && manually run test demo:
  - Without log list
  - With log list and domain providing SCTs
  - With log list and domain not providing SCTs
  All atoms are reported as expected.
Flag: com.android.org.conscrypt.flags.certificate_transparency_platform
Change-Id: Ibf93147e1014fafb44a52098f4552418b9a61850

* Refactor metrics classes

Regenerate ConscryptStatsLog based on the latest proto definition. This
class is nearly identical to the generated wrapper except for the usage
of ReflexiveStatsEvent and ReflexiveStatsLog. Updated the class
documentation with the constraints around its usage.

buildEvent methods are not used as non-existent in the wrapper. They are
marked @Deprecated but not removed because implicitly used by
MetricsTest in CTS. ReflexiveStatsEvent builder is updated to ignore any
call to writeIntArray if not supported by the platform (i.e., sdk < 32).

The Source class is updated to use the ConscryptStatsLog constants. A
new class CertificateTransparencyVerificationReason is added to
encapsulate the reason constants.

From a behaviour's perspective, this change is a no-op.

Test: atest
Bug: 319829948
Change-Id: Ie6b166530544345491347ff6bda8bbf4633e37c3

* Fix import in android/.../Platform.java