Revert git-clang-format hook and restore import ordering changes.
The new git-clang-format organizes imports in a way that is not accepted
by gms and thus causes a lot of problems in downstreaming. This cl
reverts that hooks and organises imports according to what is expected
within gms in order to avoid future conflicts. We should readd the hook
when we manage to ensure import ordering is consistent across repos.

I think I've got all import changes but let me know if I've missed any.
Some of these changes will have to be pushed into aosp afterwards.

TESTED=n/a