Use ByteArrayOutputStream to buffer data in Conscrypt's AEAD Ciphers. (#1318)
* Use ByteArrayOutputStream to buffer data in Conscrypt's AEAD Ciphers.

Currently, the input data is buffered by a byte array, which size
is increased when there is not enough space. Getting all edge-cases
right is difficult. It is better to use ByteArrayOutputStream for this.
(We create a subclass of it to avoid copying the whole array.)

Also, we remove the global variable lastGlobalMessageSize, while keeping
the optimizations it achived:
- optimize encryptions of data of the same size.
- avoid keeping a large buffer, if it subsequent encryptions
  use short message.

We avoid some unnecessary memory allocations on initialization.