Validate Ed25519 keys when deserializing them.
This changes the behavior of the deserialization: before this change,
deserializing an invalid serialization may not fail, but will fail
later when the key is used.

This change makes it fail on deserialization, which
is preferable. We don't expect users to be affected by this, because
such invalid serialization are not used.