Remove subprojects logic in favour of convention plugins.
Should be a no-op for Gradle builds and definitely a no-op
everywhere else.  I didn't manage to test signing still works
but will do at some point.

No changes in the logic of what is applied to each project and
I've kept the convention plugins in Groovy for now for ease
of review.

This is unfortunately a slightly large change but should be
straightforward to review and unlocks the path to other,
smaller, incremental gradle improvements.

Planned next steps (in no particular order):
* Move the NativeBuildInfo logic to a plugin and support ARM Linux.
* Simplify boringssl and property file logic
* Migrate to a recent Gradle
* Use recent Gradle toolchain logic to simplify native builds
* Simplify JVM task logic