Add missing evaluationDependsOn.
This was a latent bug all along.  When Android SDK is present,
android/build.gradle already contains this so openjdk was
getting it that way.  No idea why it was working on non-SDK
builds though...

Note, this is fugly, but it's what we have right now.  Once
this PR lands, I will do a smaller follow-up change which
untangles the constants dependency a bit, and when we get
to Gradle 9 we can migrate to "proper" C++ toolchains.