Remove subprojects logic in favour of convention plugins. (#1386)
* Remove subprojects logic in favour of convention plugins.

Should be a no-op for Gradle builds and definitely a no-op
everywhere else.  I didn't manage to test signing still works
but will do at some point.

No changes in the logic of what is applied to each project and
I've kept the convention plugins in Groovy for now for ease
of review.

This is unfortunately a slightly large change but should be
straightforward to review and unlocks the path to other,
smaller, incremental gradle improvements.

Planned next steps (in no particular order):
* Move the NativeBuildInfo logic to a plugin and support ARM Linux.
* Simplify boringssl and property file logic
* Migrate to a recent Gradle
* Use recent Gradle toolchain logic to simplify native builds
* Simplify JVM task logic

* Typos.

* Add missing evaluationDependsOn.

This was a latent bug all along.  When Android SDK is present,
android/build.gradle already contains this so openjdk was
getting it that way.  No idea why it was working on non-SDK
builds though...

Note, this is fugly, but it's what we have right now.  Once
this PR lands, I will do a smaller follow-up change which
untangles the constants dependency a bit, and when we get
to Gradle 9 we can migrate to "proper" C++ toolchains.

* Fix local uberjar publishing.

Missing boringssl convention plugin.

Also needs explicit mavenLocal() repository, which we should
refactor later to prevent any release accidents.

Tidied up the boringssl convention plugin a bit while I was here.