Add test-vector tests for ML-DSA and EdDSA with slightly different serializations. (#1427)
Add test-vectors with slightly different serializations. These test-vectors were generated with the same key classes, but when "writeObject" is implemented with the default implementation. For some weird reason, the serialization changes once writeObject is implemented: one byte changes (the "SC_WRITE_METHOD" bit is flipped) and an additional byte at the end is added.

Both formats are valid and all implementations of this class are able to read both of them. I want to add these tests to make sure the current implementation can read them. Later, I want to add writeObject to these keys because I want to change their implementation without breaking the serialization.